Smart Car with Arduino Mega and a Gerduino
The Arduino Mega 2560 is connected to two pairs of ultrasonic sensor and
a obstacle avoidance sensor, one in front and one in back.  The pair of
sensors are controlled by a servo each, to move sideways.  Also
connected to a bluetooth module and acts
as an i2c master.  The Gertduino sitting on top of a Raspberry Pi acts
as an i2c slave and is connected to a L293D motor shield which in turn
is connected to four motors of the car.  The Raspberry Pi on the car is
connected to a Raspberry Pi Camera which can be moved sideways or
up/down through two servos.  Also another Raspberry Pi sitting on a
table is connected to a webcam which in turn can be turned sideways or
up/down through two servos controlled by another Arduino Mega 2560.